[MOVE] move entity, guards, types folders to common folder
