+++ src/user/user.controller.ts
@@ -27,8 +27,8 @@ import {
 } from './dto/response/response.dto';
 import { KakaoSignOutReqDto } from '@src/user/dto/request/kakao-req.dto';
 import { NicknameReqDto, UidReqDto } from './dto/request/user-req.dto';
-import { HoodieAuthTokens } from '../types/user';
-import { AuthGuard } from '@src/guards/auth.guard';
+import { HoodieAuthTokens } from '../common/types/user';
+import { AuthGuard } from '@src/common/guards/auth.guard';
 
 @ApiTags('user')
 @Controller('user')
